requirements.txt
Flask
python-dotenv
PyQt6
requests
supabase
openai
openai-agents